make custom pagination items as a component
